Faster and more accurate Hellinger distance
Calculate the Hellinger distance using the Bhattacharyya coefficient. This reduces the number of sqrt calculations while being mathematically equivalent.

Also, normalize the histogram features such that they can be interpreted as an estimate of the probability density function of the respective channel sampled in a given neighborhood. This finally yields an more appropriate usage of the Hellinger distance.